requirements.txt
flask
flask-cors
pymongo
python-dotenv
flask-socketio
gevent
gevent-websocket
paho-mqtt
